LayoutTests/fast/dom/document-collection-idempotence.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("This test verifies that the HTMLCollection getters on document (excluding .all) returns the same object when called repeatedly.");

var collections = [ "images", "embeds", "plugins", "applets", "links", "forms", "anchors", "scripts" ];

shouldBe("document.all === document.all", "false");

for (i = 0; i < collections.length; ++i) {
    var collection = collections[i];
    shouldBe("document." + collection + " === document." + collection, "true");
}

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
