LayoutTests/fast/dom/document-collection-idempotence-expected.txt

This test verifies that the HTMLCollection getters on document return the same object when called repeatedly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.all === document.all is true
PASS document.images === document.images is true
PASS document.embeds === document.embeds is true
PASS document.plugins === document.plugins is true
PASS document.applets === document.applets is true
PASS document.links === document.links is true
PASS document.forms === document.forms is true
PASS document.anchors === document.anchors is true
PASS document.scripts === document.scripts is true
PASS successfullyParsed is true

TEST COMPLETE
